Allow running simulate_queue_worker from a Rake task
